fix code style and log
