+++ dispatchers/dispatcher_base.py
@@ -3,22 +3,24 @@ import random
 from intrabuildingtransport.mansion.utils import EPSILON, HUGE
 from intrabuildingtransport.mansion.utils import MansionAttribute
 
+
 class DispatcherBase(object):
-  '''
-  A basic demonstration of dispatcher
-  A dispatcher must provide policy and feedback function
-  The policy function receives MansionState and output ElevatorAction Lists
-  The feedback function receives reward
-  '''
-  def link_mansion(self, mansion_attr):
-    assert isinstance(mansion_attr, MansionAttribute)
-    self._mansion_attr = mansion_attr
+    '''
+    A basic demonstration of dispatcher
+    A dispatcher must provide policy and feedback function
+    The policy function receives MansionState and output ElevatorAction Lists
+    The feedback function receives reward
+    '''
 
-  def load_settings(self):
-    pass
+    def link_mansion(self, mansion_attr):
+        assert isinstance(mansion_attr, MansionAttribute)
+        self._mansion_attr = mansion_attr
 
-  def feedback(self, state, action, r):
-    pass
+    def load_settings(self):
+        pass
 
-  def policy(self, state):
-    raise NotImplementedError()
+    def feedback(self, state, action, r):
+        pass
+
+    def policy(self, state):
+        raise NotImplementedError()
